EDGEAI-364 Refactored fpsUpdate function into stats
