+++ src/js/grid.js
@@ -1,6 +1,6 @@
 import * as THREE from './three.js'
 import { OrbitControls } from './OrbitControls.js'
-import Stats from './Stats.js'
+import Stats, { fpsUpdate } from './Stats.js'
 import pcdStream from './pcd.js'
 import { parseNumbersInObject } from './parseNumbersInObject.js';
 import { grid_set_radarpoints, init_grid } from './grid_render.js'
@@ -27,41 +27,6 @@ const radarPanel = stats.addPanel(new Stats.Panel('radarFPS', '#ff4', '#220'));
 stats.dom.style.cssText = "position: absolute; top: 0px; right: 0px; opacity: 0.9; z-index: 10000;";
 stats.showPanel([])
 document.querySelector('main').appendChild(stats.dom);
-
-
-function fpsUpdate(panel, max) {
-    if (!max) {
-        max = 40
-    }
-    let fpsInd = 0
-    let timeBetweenUpdates = []
-    let lastUpdateTime = 0
-    let stablized = false
-    let firstUpdate = 0
-    return () => {
-        if (!lastUpdateTime) {
-            lastUpdateTime = performance.now()
-            firstUpdate = lastUpdateTime
-            return
-        }
-        const curr = performance.now()
-        if (timeBetweenUpdates.length < 10) {
-            timeBetweenUpdates.push(curr - lastUpdateTime)
-            lastUpdateTime = curr
-            return
-        }
-        timeBetweenUpdates[fpsInd] = curr - lastUpdateTime
-        fpsInd = (fpsInd + 1) % timeBetweenUpdates.length
-        if (stablized) {
-            const avg_fps = 1000 / timeBetweenUpdates.reduce((a, b) => a + b, 0) * timeBetweenUpdates.length
-            panel.update(avg_fps, max)
-        } else if (curr - firstUpdate > 2000) {
-            // has been 2s since first update
-            stablized = true
-        }
-        lastUpdateTime = curr
-    }
-}
 
 const loader = new THREE.FileLoader();
